requirements.txt
mysql-connector-python==9.1.0
pwinput==1.0.3
pycryptodome==3.21.0
pyotp==2.9.0
pyperclip==1.9.0
python-dotenv==1.0.1
qrcode==8.0
